+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2022-06-14
+- Change item display name
+- Modular search argument creation
+- Use url parameter setting search content
+
+## [0.1.3] - 2022-06-07
+- Add searching option
+
+## [0.1.2] - 2022-05-31
+### [Added]
+- Can use the specify url to search FHIR data
+
+## [0.1.1] - 2022-05-24
+### [Added]
+- List the all patient detail on page
+
 ## [0.1.0] - 2022-05-12
 ### [Added]
 - Get patient data from the FHIR server